Fix tests for new openapi
